web: split maxConns in api and backend
previously, there was a single `maxConns` configuration flag that'd be
used to set a max number of conn for both api and backend. as this flag
does not exist anymore, it was removed.

given that it's usually quite unequal how these pools gets used, in 6.0,
these configs are separated.

ps.: note that there's actually a third pool - gc - which is hardcoded
to 5.

Signed-off-by: Ciro S. Costa <cscosta@pivotal.io>